Restructure scraper modules
