--- JobScraper/justjoin-scraper/host.json
@@ -1,17 +1,0 @@
-{
-  "version": "2.0",
-
-  // <— add this block
-  "extensionBundle": {
-    "id": "Microsoft.Azure.Functions.ExtensionBundle",
-    "version": "[3.*, 4.0.0)"
-  },
-
-  "logging": {
-    "applicationInsights": {
-      "samplingSettings": {
-        "isEnabled": true
-      }
-    }
-  }
-}
