+++ JobScraper/pracuj-scraper/__init__.py
@@ -1,0 +1,6 @@
+import azure.functions as func
+from .pracuj_scraper import PracujScraper
+
+def main(timer: func.TimerRequest):
+    scraper = PracujScraper()
+    scraper.scrape()
