+++ JobScraper/pracuj-scraper/function.json
@@ -1,0 +1,11 @@
+{
+  "bindings": [
+    {
+      "name": "timer",
+      "type": "timerTrigger",
+      "direction": "in",
+      "schedule": "0 0 9 * * *",
+      "runOnStartup": true
+    }
+  ]
+}
